requirements.txt
numpy==1.15.0
Pillow==3.1.2
scipy==0.19.0
tensorflow=1.7.0